+++ src/components/Header/header.module.scss
@@ -20,3 +20,81 @@
         cursor: pointer;
     }
 }
+
+.subheader {
+    display: flex;
+    height: 56px;
+    background-color: #373F68;
+    color: #F2F4FE;
+    align-items: center;
+    justify-content: space-between;
+    padding: 0 24px;
+    &_selectedOption {
+        display: flex;
+        flex-direction: row;
+        position: relative;
+    }
+    &_selector {
+        cursor: pointer;
+    }
+    &_button {
+        cursor: pointer;
+        height: 40px;
+        width: 134px;
+        background-color: #AD1FEA;
+        border-radius: 10px;
+        text-align: center;
+        //this will be modified once the button component is created... it will be an element
+        h4 {
+            font-size: 13px;
+            color: #F2F4FE;
+        }
+    }
+}
+
+.roadmapHeader {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    height: 100px;
+    background-color: #373F68;
+    color: #FFFFFF;
+    padding: 24px 26px;
+    box-sizing: border-box;
+    &_backButton {
+        display: flex;
+        align-items: center;
+        cursor: pointer;
+        padding-bottom: 3px;
+    }
+
+    &_addButton {
+        cursor: pointer;
+        height: 40px;
+        width: 134px;
+        background-color: #AD1FEA;
+        border-radius: 10px;
+        text-align: center;
+        h4 {
+            color: #F2F4FE;
+            font-size: 13px;
+        }
+    }
+
+    &_icon {
+        padding-right: 16px;
+        path {
+            stroke:#CDD2EE;
+        }
+       
+    }
+}
+
+
+@media screen and (max-width: 340px) {
+    .subheader {
+        &_button {
+            width: 114px;
+        }
+    }  
+}
