--- src/components/Subheader/subheader.module.scss
@@ -1,40 +1,0 @@
-.subheader {
-    display: flex;
-    height: 56px;
-    background-color: #373F68;
-    color: #F2F4FE;
-    align-items: center;
-    justify-content: space-between;
-    padding: 0 24px;
-    &_selectedOption {
-        display: flex;
-        flex-direction: row;
-        position: relative;
-    }
-    &_selector {
-        cursor: pointer;
-    }
-
-    &_button {
-        cursor: pointer;
-        height: 40px;
-        width: 134px;
-        background-color: #AD1FEA;
-        border-radius: 10px;
-        text-align: center;
-        //this will be modified once the button component is created... it will be an element
-        h4 {
-            font-size: 13px;
-        }
-        
-
-    }
-}
-
-@media screen and (max-width: 340px) {
-    .subheader {
-        &_button {
-            width: 114px;
-        }
-    }  
-}
